backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-telegram-bot==21.6
lightgbm==4.5.0
torch==2.4.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
ccxt==4.3.89
apscheduler==3.10.4
websockets==13.1
httpx==0.27.2
shap==0.46.0
scipy==1.14.1
ta==0.11.0
